Extract overlays from app.py
